fix: harmonize case surveillance tracker
